% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.seqICP.R
\name{summary.seqICP}
\alias{summary.seqICP}
\title{summary function}
\usage{
\method{summary}{seqICP}(object, show.past = TRUE, ...)
}
\arguments{
\item{object}{object of class 'seqICP'.}

\item{show.past}{'TRUE' if lagged variables should also be shown.}

\item{...}{additional arguments affecting the summary produced.}
}
\description{
Summary functions for 'seqICP' objects.
}
\author{
Niklas Pfister and Jonas Peters
}
