% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.seqICPnl.R
\name{summary.seqICPnl}
\alias{summary.seqICPnl}
\title{summary function}
\usage{
\method{summary}{seqICPnl}(object, show.past = TRUE, ...)
}
\arguments{
\item{object}{object of class 'seqICPnl'.}

\item{show.past}{'TRUE' if lagged variables should also be shown.}

\item{...}{additional arguments affecting the summary produced.}
}
\description{
Summary functions for 'seqICPnl' objects.
}
\author{
Niklas Pfister and Jonas Peters
}
